the same problem of using integer division for float values happened a few more times
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2733 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/packaging.c
@@ -1029,7 +1029,7 @@ halfword vpackage(halfword p, scaled h, int m, scaled l)
         glue_order(r) = o;
         glue_sign(r) = stretching;
         if (total_stretch[o] != 0) {
-            glue_set(r) = unfloat(x / total_stretch[o]);
+	  glue_set(r) = unfloat((double)x / total_stretch[o]);
         } else {
             glue_sign(r) = normal;
             set_glue_ratio_zero(glue_set(r));   /* there's nothing to stretch */
@@ -1070,7 +1070,7 @@ halfword vpackage(halfword p, scaled h, int m, scaled l)
         glue_order(r) = o;
         glue_sign(r) = shrinking;
         if (total_shrink[o] != 0) {
-            glue_set(r) = unfloat((-x) / total_shrink[o]);
+	  glue_set(r) = unfloat((double)(-x) / total_shrink[o]);
         } else {
             glue_sign(r) = normal;
             set_glue_ratio_zero(glue_set(r));   /* there's nothing to shrink */
